Added Timer Controller, Fixed overlap issue
