+++ app/javascript/controllers/index.js
@@ -12,3 +12,6 @@ application.register("hello", HelloController)
 
 import MapController from "./map_controller"
 application.register("map", MapController)
+
+import TimerController from "./timer_controller"
+application.register("timer", TimerController)
