+++ app/views/trips/show.html.erb
@@ -2,7 +2,9 @@
 <% zone = ActiveSupport::TimeZone.new("Melbourne") %>
 
 <div class="m-5 text-center text-3xl"><strong>Time left on your park: </strong></div>
-<div id="app" class ="flex justify-center items-center" ></div>
+<div data-controller="timer" id="app" class ="flex justify-center items-center" >
+
+</div>
 <div class="container mx-auto text-center mt-5 p-4 ">
 <p><strong> Address: </strong> <%= @trip.parking_bay.address %> </p>
 <p><strong> Bay Number: </strong> <%= @trip.parking_bay.id %> </p>
